Projects/discord_py/requirements.txt
discord.py
python-dotenv
async
asyncio
